Refactor out generic setup and teardown logic for tests
