+++ tests/test_game_lobby_repository.py
@@ -2,12 +2,87 @@ from datetime import datetime, timedelta
 from time import gmtime, strftime
 
 import pytest
+import pytest_asyncio
+from dotenv import load_dotenv
 from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy.orm import sessionmaker
 
+from repository.db_config import Base
 from repository.lobby_repo import LobbyPostgresRepository
+from repository.table.game_lobby_tables import (GameModel, GuildModel,
+                                                LobbyModel, MemberModel)
+
+load_dotenv()
 
 
-class TestRepository:
+@pytest_asyncio.fixture(scope="session", autouse=True)
+async def init_database(engine):
+    """Test specific: Seed database with example entries"""
+    async with engine.begin() as session:
+        await session.run_sync(Base.metadata.create_all)
+
+    async_session = sessionmaker(engine, expire_on_commit=True, class_=AsyncSession)
+
+    async with async_session() as session:
+        member = MemberModel(id=123)
+        session.add(member)
+
+        member2 = MemberModel(id=321)
+        session.add(member2)
+
+        guild = GuildModel(
+            id=1,
+            name="test",
+        )
+
+        session.add(guild)
+
+        await session.flush()
+
+        game = GameModel(
+            guild_id=guild.id,
+            name="test",
+            max_size=5,
+        )
+
+        game2 = GameModel(
+            guild_id=guild.id,
+            name="test2",
+            max_size=2,
+        )
+
+        session.add(game)
+        session.add(game2)
+
+        await session.flush()
+
+        lobby = LobbyModel(
+            created_datetime=datetime.now(),
+            original_channel_id=12,
+            lobby_channel_id=34,
+            control_panel_message_id=56,
+            description="test",
+            embed_message_id=78,
+            queue_message_id=91,
+            game_id=game.id,
+            guild_id=guild.id,
+            game_size=5,
+            last_promotion_message_id=None,
+            last_promotion_datetime=None,
+            history_thread_id=12,
+            is_locked=False,
+            owner_id=member.id,
+        )
+
+        lobby.members.append(member)
+        lobby.queue_members.append(member2)
+
+        session.add(lobby)
+
+        await session.commit()
+
+
+class TestGameLobbyRepository:
 
     """QUEUE MEMBER TESTS"""
 
@@ -414,7 +489,7 @@ class TestRepository:
             user_id=321,
             game_id=1,
             max_size=5,
-            description=""
+            description="",
         )
         result = await lobbydal.get_lobbies_count()
         assert result == 2
