+++ tests/test_reminder_repository.py
@@ -1,0 +1,2 @@
+class TestReminderRepository:
+    pass
